requirements.txt
numpy
scipy
pandas
jax
jaxlib
flax
pyscf
tensorcircuit-ng[qiskit]
openfermion
pylatexenc
noisyopt
renormalizer==0.0.10
